toFloat

+++ src/main/scala/com/yogpc/qp/machines/item/GuiQuarryLevel.scala
@@ -56,7 +56,7 @@ class GuiQuarryLevel(c: ContainerQuarryLevel, inv: PlayerInventory, t: ITextComp
     val s: String = tileName
     this.font.drawString(s, this.xSize / 2 - this.font.getStringWidth(s) / 2, 6, 0x404040)
     this.font.drawString(I18n.format(TranslationKeys.CONTAINER_INVENTORY), 8, this.ySize - 96 + 2, 0x404040)
-    this.font.drawString(lA.getYLevel(tile).toString, this.xSize / 2 - this.font.getStringWidth(lA.getYLevel(tile).toString) / 2, tp + 23, 0x404040)
+    this.font.drawString(lA.getYLevel(tile).toString, (this.xSize / 2 - this.font.getStringWidth(lA.getYLevel(tile).toString) / 2).toFloat, tp.toFloat + 23, 0x404040)
   }
 
   override def actionPerformed(button: IHandleButton.Button): Unit = {
